put the thread class in separate file
